test: consolidate signal canceller integration coverage
